Further work integrating network system with message channels.
