+++ gestalt-core/src/main.rs
@@ -34,7 +34,7 @@ use common::{identity::{do_keys_need_generating, does_private_key_need_passphras
 use std::collections::{HashSet, HashMap};
 use tokio::sync::mpsc;
 
-use crate::{net::{PREPROTCOL_PORT, NetworkRole, preprotocol::{launch_preprotocol_listener, preprotocol_connect_to_server}, GESTALT_PORT, run_network_system, LaminarConfig, TypedNetMsgReceiver, net_channels::{NetSendChannel, net_msg_channel}, NetMsg}, common::{identity::generate_local_keys}, message_types::{voxel::{VoxelChangeAnnounce, VoxelChangeRequest}, JoinDefaultEntry, JoinAnnounce}, message::{START_QUIT, QuitReceiver}};
+use crate::{net::{PREPROTCOL_PORT, NetworkRole, preprotocol::{launch_preprotocol_listener, preprotocol_connect_to_server}, GESTALT_PORT, run_network_system, LaminarConfig, net_channels::{NetSendChannel, net_send_channel, net_recv_channel}, NetMsg}, common::{identity::generate_local_keys}, message_types::{voxel::{VoxelChangeAnnounce, VoxelChangeRequest}, JoinDefaultEntry, JoinAnnounce}, message::{START_QUIT, QuitReceiver}};
 
 pub const ENGINE_VERSION: Version = version!(0,0,1);
 
@@ -281,20 +281,7 @@ fn main() {
     };
 
     info!("Setting up channels.");
-    
-    let (client_voxel_sender_from_server, client_voxel_receiver_from_server) = tokio::sync::broadcast::channel(4096);
-    let client_voxel_receiver_from_server: TypedNetMsgReceiver<VoxelChangeAnnounce> = TypedNetMsgReceiver::new(client_voxel_receiver_from_server);
 
-    let (server_voxel_sender_from_client, server_voxel_receiver_from_client) = tokio::sync::broadcast::channel(4096);
-    let mut server_voxel_receiver_from_client: TypedNetMsgReceiver<VoxelChangeRequest> = TypedNetMsgReceiver::new(server_voxel_receiver_from_client);
-
-
-    let (client_join_sender_from_server, client_join_receiver_from_server) = tokio::sync::broadcast::channel(4096);
-    let mut client_join_receiver_from_server: TypedNetMsgReceiver<JoinAnnounce> = TypedNetMsgReceiver::new(client_join_receiver_from_server);
-
-    let (server_join_sender_from_client, server_join_receiver_from_client) = tokio::sync::broadcast::channel(4096);
-    let mut server_join_receiver_from_client: TypedNetMsgReceiver<JoinDefaultEntry> = TypedNetMsgReceiver::new(server_join_receiver_from_client);
-    
     let mut laminar_config = LaminarConfig::default();
     laminar_config.heartbeat_interval = Some(Duration::from_secs(1));
 
@@ -322,8 +309,7 @@ fn main() {
             run_network_system(NetworkRole::Server,
                 udp_address, 
                 connect_receiver,
-                keys.clone(), 
-                HashMap::from([(VoxelChangeRequest::net_msg_id(), server_voxel_sender_from_client), (JoinDefaultEntry::net_msg_id(), server_join_sender_from_client)]),
+                keys.clone(),
                 laminar_config,
                 Duration::from_millis(25))
         );
@@ -346,20 +332,22 @@ fn main() {
         let mut total_changes: Vec<VoxelChangeAnnounce> = Vec::new();
         async_runtime.block_on(async move { 
             let mut quit_receiver = QuitReceiver::new();
+            let mut server_voxel_receiver_from_client = net_recv_channel::subscribe::<VoxelChangeRequest>().unwrap();
+            let mut server_join_receiver_from_client = net_recv_channel::subscribe::<JoinDefaultEntry>().unwrap();
             loop {
                 tokio::select! { 
-                    voxel_events_maybe = server_voxel_receiver_from_client.recv() => { 
+                    voxel_events_maybe = server_voxel_receiver_from_client.recv_wait() => { 
                         if let Ok(voxel_events) = voxel_events_maybe { 
                             for (ident, event) in voxel_events { 
                                 //world_space.set(event.pos, event.new_tile).unwrap();
                                 info!("Received {:?} from {}", &event, ident.to_base64());
                                 let announce: VoxelChangeAnnounce = event.into();
-                                net_msg_channel::send_to_all_except(announce.clone(), &ident).unwrap();
+                                net_send_channel::send_to_all_except(announce.clone(), &ident).unwrap();
                                 total_changes.push(announce);
                             }
                         }
                     }
-                    join_event_maybe = server_join_receiver_from_client.recv() => { 
+                    join_event_maybe = server_join_receiver_from_client.recv_wait() => { 
                         if let Ok(events) = join_event_maybe { 
                             for (ident, event) in events {
                                 info!("User {} has joined with display name {}", ident.to_base64(), &event.display_name);
@@ -367,9 +355,9 @@ fn main() {
                                     display_name: event.display_name, 
                                     identity: ident,
                                 };
-                                net_msg_channel::send_to_all_except(announce.clone(), &ident).unwrap();
+                                net_send_channel::send_to_all_except(announce.clone(), &ident).unwrap();
                                 info!("Sending all previous changes to the newly-joined user.");
-                                net_msg_channel::send_multi_to(total_changes.clone(), &ident).unwrap();
+                                net_send_channel::send_multi_to(total_changes.clone(), &ident).unwrap();
                             }
                         }
                     }
@@ -396,7 +384,6 @@ fn main() {
             run_network_system( NetworkRole::Client,  address, 
                 connect_receiver,
                 keys.clone(), 
-                HashMap::from([(VoxelChangeAnnounce::net_msg_id(), client_voxel_sender_from_server), (JoinAnnounce::net_msg_id(), client_join_sender_from_server)]),
                 laminar_config,
                 Duration::from_millis(25))
         );
@@ -407,11 +394,14 @@ fn main() {
 
         std::thread::sleep(Duration::from_millis(50));
                 
-        let voxel_event_sender: NetSendChannel<VoxelChangeRequest> = net_msg_channel::subscribe_sender(&server_identity).unwrap();
+        let voxel_event_sender: NetSendChannel<VoxelChangeRequest> = net_send_channel::subscribe_sender(&server_identity).unwrap();
         
+        let mut client_join_receiver_from_server = net_recv_channel::subscribe::<JoinAnnounce>().unwrap();
+        let client_voxel_receiver_from_server = net_recv_channel::subscribe::<VoxelChangeAnnounce>().unwrap();
+
         async_runtime.spawn( async move { 
             loop { 
-                match client_join_receiver_from_server.recv().await { 
+                match client_join_receiver_from_server.recv_wait().await { 
                     Ok(join_msgs) => {
                         for (_server_ident, JoinAnnounce{identity, display_name }) in join_msgs { 
                             info!("Peer {} joined with display name {}", identity.to_base64(), &display_name);
@@ -441,6 +431,8 @@ fn main() {
     else {
         let (voxel_event_sender, mut voxel_event_receiver) = tokio::sync::broadcast::channel(4096); 
         let voxel_event_sender = NetSendChannel::new(voxel_event_sender); 
+
+        let client_voxel_receiver_from_server = net_recv_channel::subscribe::<VoxelChangeAnnounce>().unwrap();
 
         async_runtime.spawn( async move {
             loop {
